+++ src/App.jsx
@@ -1,15 +1,11 @@
 import React, { useEffect } from 'react'
 import './App.css'
 import './index.css'
-
 import Content from './components/Layout/Content/Content'
 import PageHeader from './components/Layout/PageHeader/PageHeader'
 import PublishModal from './components/PublishModal/PublishModal'
-import LaunchModal from './components/LaunchModal/LaunchModal'
-import LaunchImageModal from './components/LaunchModal/LaunchImageModal'
 import UploadGeojsonModal from './components/UploadGeojsonModal/UploadGeojsonModal'
 import SystemMessage from './components/SystemMessage/SystemMessage'
-
 import { GetCollectionsService } from './services/get-collections-service'
 import { LoadConfigIntoStateService } from './services/get-config-service'
 import { useSelector } from 'react-redux'
@@ -17,12 +13,6 @@ import { useSelector } from 'react-redux'
 function App() {
   const _showPublishModal = useSelector(
     (state) => state.mainSlice.showPublishModal
-  )
-  const _showLaunchModal = useSelector(
-    (state) => state.mainSlice.showLaunchModal
-  )
-  const _showLaunchImageModal = useSelector(
-    (state) => state.mainSlice.showLaunchImageModal
   )
   const _showUploadGeojsonModal = useSelector(
     (state) => state.mainSlice.showUploadGeojsonModal
@@ -49,8 +39,6 @@ function App() {
           <PageHeader></PageHeader>
           <Content></Content>
           {_showPublishModal ? <PublishModal /> : null}
-          {_showLaunchModal ? <LaunchModal /> : null}
-          {_showLaunchImageModal ? <LaunchImageModal /> : null}
           {_showUploadGeojsonModal ? (
             <UploadGeojsonModal></UploadGeojsonModal>
           ) : null}